Migrate datetime types to icu_calendar crate
